requirements.txt
# Product Matching System - Complete Requirements
#
# PYTHON VERSION: 3.12 (Required for all platforms)
# This ensures compatibility with AMD ROCm and simplifies deployment
#
# QUICK START:
# 1. Install Python 3.12 from https://www.python.org/downloads/
# 2. Install requirements: pip install -r requirements.txt
# 3. For GPU acceleration: python gpu/setup_gpu.py (optional but recommended)
#
# GPU SUPPORT (Automatic Detection):
# - AMD ROCm (Windows): Requires Python 3.12 ✓
#   * ROCm 7.1.1 (Nov 2024): PyTorch 2.9, SDPA improvements, AOTriton 0.11b
#   * ROCm 6.4.4: PyTorch 2.8, stable and well-tested
#   * Both versions work with sentence-transformers < 3.0.0 restriction
# - NVIDIA CUDA: Works with Python 3.12 ✓
# - Apple Silicon (macOS): Works with Python 3.12 ✓
# - CPU fallback: Works with Python 3.12 ✓
#
# Using Python 3.12 ensures maximum compatibility across all GPU types

# Web Framework (Required for Flask app)
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug>=3.0.0

# Core Scientific Computing
numpy>=1.26.2
scipy>=1.11.3

# Image Processing
opencv-python>=4.8.1.78
scikit-image>=0.22.0
Pillow>=10.1.0

# Desktop Wrapper (optional - for standalone app)
pywebview==4.4.1

# Deep Learning (GPU users: install via gpu/setup_gpu.py for optimal GPU support)
# PyTorch - Version depends on GPU:
#   - AMD ROCm (Windows): 2.8.0 (ROCm 6.4.4) or 2.9.0 (ROCm 7.1.1) - Python 3.12 only
#   - NVIDIA CUDA: 2.x with CUDA 12.4
#   - Apple Silicon (macOS): 2.x with MPS
#   - CPU: 2.x CPU-only
torch>=2.0.0

# CLIP Model for Visual Matching
# CRITICAL: sentence-transformers MUST BE < 3.0.0 for AMD ROCm compatibility
# Version 3.0+ requires torch.distributed.is_initialized() which is missing in ROCm Windows builds
# Tested with ROCm 6.4.4 and 7.1.1 (PyTorch 2.9) - restriction still applies
# DO NOT UPGRADE to 3.0+ or AMD GPU support will break!
# Note: sentence-transformers will auto-install transformers >= 4.55.5 (required for ROCm)
sentence-transformers>=2.7.0,<3.0.0

# Intel GPU Acceleration (optional - for Intel Arc/Iris GPUs)
# Uncomment to enable Intel GPU support:
# intel-extension-for-pytorch>=2.1.0
# Provides 3-5x speedup on Intel GPUs vs CPU
# Works on: Intel Arc, Intel Iris Xe, Intel UHD Graphics (11th gen+)
# Installation: pip install intel-extension-for-pytorch
# Note: Automatically detected and used if installed, gracefully falls back to CPU if not

# Testing (optional - for development)
pytest>=7.0.0
pytest-timeout>=2.1.0
